Nice! Write the "namespace" thing and, also, strings can be compared now
The syntax is:

    "namespace" NAME stmt

Example:

    namespace foo {
      my bar = "one";
      our baz = "two";
    }

This would create a private variable "bar" and a public variable "baz" inside
the "foo" namespace.

NOTE: if the namespace "foo" didn't exist before, it would be created.
If it did exist before, it would be "reopened" and the variables/functions would
be inserted into it.