Start some work on those so-called accumulators
Although, so far it's only a working title. I should probably figure out a
better name for them.

Basically, accumulators (which semanticaly are defined as: %[0-9]+) are
responsible for a number of things, they for example would be arguments to a
function, or args passed to the script, and so on.